+++ src/built_in/execute_builtin_function.c
@@ -6,7 +6,7 @@
 /*   By: Philip <juli@student.42london.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 01:59:40 by Philip            #+#    #+#             */
-/*   Updated: 2024/04/17 13:06:42 by Philip           ###   ########.fr       */
+/*   Updated: 2024/04/17 17:14:05 by Philip           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,8 @@
 #include "../pipes/t_pipes.h"
 
 /**
- * @brief Executes the built-in function.
+ * @brief Executes the built-in function. This function does not exit the
+ *        process or frees any dynamically allocated memory.
  * 
  * @param argv arguments for the function.
  * @param env a pointer to the pointer to the environment variables.
@@ -38,6 +39,8 @@ extern int	execute_builtin_function(char **argv, t_env **env, t_cmd_list *cmds,
 		exit_status = builtin_echo(argv);
 	else if (ft_strncmp(argv[0], "exit", 5) == 0)
 		builtin_exit(argv, *env, cmds, pipes);
+	else if (ft_strncmp(argv[0], "unset", 6) == 0)
+		builtin_unset(argv, env);
 	else if (ft_strncmp(argv[0], "export", 7) == 0)
 		exit_status = builtin_export(argv, env);
 	return (exit_status);
