--- src/heredoc.h
@@ -1,18 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   heredoc.h                                          :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: Philip <juli@student.42london.com>         +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/11 18:21:32 by Philip            #+#    #+#             */
-/*   Updated: 2024/04/11 18:22:15 by Philip           ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef HEREDOC_H
-# define HEREDOC_H
-
-# define HEREDOC_FILE "./.heredoc_temp"
-
-#endif
